+++ changelog.md
@@ -1,3 +1,13 @@
+## [v1.0.1 (2025-2-28)]
+
+- Added support for multiple HTTP methods in Router class
+- Added Route Caching in Router class
+- Added DocBlock to Core files
+- Fixed namespace resolution issues in Router class
+- Improved session system for better security and performance
+- Improved Router class for better performance and flexibility
+- Improved Base class for better performance and flexibility
+
 ## [v1.0.0 (2023-11-21)]
 
-Initial release of Phuse 1, based on collective and collaborative framework named 'Orceztra'(discontinued personal project).
+ - Initial release of Phuse 1, based on collective and collaborative framework named 'Orceztra'(discontinued personal project).